Updated repositories

+++ src/main/java/com/super_famicom_coder/songr/AlbumRepository.java
@@ -4,4 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlbumRepository extends CrudRepository<Album, Long> {}
+public interface AlbumRepository extends CrudRepository<Album, Integer> {
+
+}
